requirements.txt
streamlit
pandas
pyodbc
python-dotenv
openai
